Rubocop autocorrect for models

+++ app/models/tagging.rb
@@ -2,7 +2,7 @@ class Tagging < ActiveRecord::Base
   belongs_to :tag
   belongs_to :taggable, polymorphic: true
 
-  belongs_to :verse,  foreign_key: 'taggable_id'
+  belongs_to :verse, foreign_key: 'taggable_id'
 
   after_destroy :destroy_tag_if_unused
 
@@ -10,9 +10,7 @@ class Tagging < ActiveRecord::Base
 
   def destroy_tag_if_unused
     if Tag.destroy_unused
-      if tag.taggings.count.zero?
-        tag.destroy
-      end
+      tag.destroy if tag.taggings.count.zero?
     end
   end
 end
